+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - kube-proxy: simplify proxy-mode configuration
 - Make kubernetes services network route ipvs compatible
 - Packages: install `ipvsadm` and `ipset` for IPVS proxy mode
+- kube-proxy: split attributes to global and ipvs-related
 
 # 1.11.3 (11.07.2018)
 - CNI plugins: bump to 0.7.1